# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Fides Gate crawler verification and accountability API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Summary {
  totalRequests: number;
  verified: number;
  fallback: number;
  canaryDetections: number;
  activeCrawlers: number;
  trustScore: number;
  blocked: number;
}

export interface RequestEvent {
  id: string;
  timestamp: string;
  crawlerId: string;
  crawlerName: string;
  intent: string;
  identity: string;
  tier: string;
  status: number;
  evidenceHash: string;
  trustScore: number;
  action: string;
  location: string;
}

export interface LedgerEvent {
  id: string;
  event: string;
  crawler: string;
  publisher: string;
  intent: string;
  timestamp: string;
  action: string;
  violation: string;
  hash: string;
  previousHash: string;
  status: string;
}

export interface Policy {
  intent: string;
  label: string;
  action: string;
  description: string;
  enabled: boolean;
  tone: string;
}

export interface Crawler {
  id: string;
  name: string;
  key: string;
  status: string;
  purpose: string;
  trustScore: number;
  requests: number;
  lastSeen: string;
  accent: string;
}

export interface CanaryStatus {
  status: string;
  crawler: string;
  session: string;
  confidence: number;
  matchedSignals: string[];
}

export interface Dashboard {
  summary: Summary;
  requests: RequestEvent[];
  ledger: LedgerEvent[];
  policies: Policy[];
  crawlers: Crawler[];
  canary: CanaryStatus;
}

export interface PolicyUpdate {
  policies: Policy[];
}

export interface ChartPoint {
  label: string;
  value: number;
  verified?: number;
  invalid?: number;
}

export interface LeaderboardEntry {
  name: string;
  score: number;
  requests: number;
  violations: number;
  status: string;
}

export interface Analytics {
  volume: ChartPoint[];
  intent: ChartPoint[];
  hourly: ChartPoint[];
  trustLeaderboard: LeaderboardEntry[];
  violations: ChartPoint[];
}

export interface Evidence {
  hash: string;
  contentHash: string;
  timestamp: string;
  crawler: string;
  publisher: string;
}

export interface GateStep {
  label: string;
  detail: string;
  status: string;
}

export interface SimulationResult {
  request: RequestEvent;
  message: string;
  payload: string;
  canary: CanaryStatus;
  evidence: Evidence;
  steps: GateStep[];
}

export interface CanaryScanInput {
  text: string;
}

export interface CanaryScanResult {
  detected: boolean;
  status: string;
  crawler: string;
  session: string;
  confidence: number;
  matchedSignals: string[];
  evidence: Evidence;
}

export interface DemoReset {
  status: string;
  message: string;
  dashboard: Dashboard;
}