# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Fides Gate crawler verification and accountability API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /policies:
    get:
      operationId: getPolicies
      tags: [policy]
      summary: Get publisher policies
      responses:
        "200":
          description: Publisher policies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Policy"
    patch:
      operationId: updatePolicies
      tags: [policy]
      summary: Update publisher policies
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PolicyUpdate"
      responses:
        "200":
          description: Updated publisher policies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Policy"
...
  /canary/scan:
    post:
      operationId: scanCanary
      tags: [canary]
      summary: Scan text for known semantic canaries
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CanaryScanInput"
      responses:
        "200":
          description: Canary scan result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CanaryScanResult"
components:
  schemas:
...
    Policy:
      type: object
      required: [intent, label, action, description, enabled, tone]
      properties:
        intent: { type: string }
        label: { type: string }
        action: { type: string }
        description: { type: string }
        enabled: { type: boolean }
        tone: { type: string }
    PolicyUpdate:
      type: object
      required: [policies]
      properties:
        policies:
          type: array
          items:
            $ref: "#/components/schemas/Policy"
...
    Evidence:
      type: object
      required: [hash, contentHash, timestamp, crawler, publisher]
      properties:
        hash: { type: string }
        contentHash: { type: string }
        timestamp: { type: string }
        crawler: { type: string }
        publisher: { type: string }
    CanaryScanInput:
      type: object
      required: [text]
      properties:
        text: { type: string }
    CanaryScanResult:
      type: object
      required: [detected, status, crawler, session, confidence, matchedSignals, evidence]
      properties:
        detected: { type: boolean }
        status: { type: string }
        crawler: { type: string }
        session: { type: string }
        confidence: { type: number }
        matchedSignals:
          type: array
          items: { type: string }
        evidence:
          $ref: "#/components/schemas/Evidence"
...