# lib/api-zod/src/generated/types/canaryScanInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Fides Gate crawler verification and accountability API
 * OpenAPI spec version: 0.1.0
 */

export interface CanaryScanInput {
  text: string;
}