# lib/api-zod/src/generated/types/canaryScanResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Fides Gate crawler verification and accountability API
 * OpenAPI spec version: 0.1.0
 */
import type { Evidence } from './evidence';

export interface CanaryScanResult {
  detected: boolean;
  status: string;
  crawler: string;
  session: string;
  confidence: number;
  matchedSignals: string[];
  evidence: Evidence;
}