# lib/api-zod/src/generated/types/policyUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Fides Gate crawler verification and accountability API
 * OpenAPI spec version: 0.1.0
 */
import type { Policy } from './policy';

export interface PolicyUpdate {
  policies: Policy[];
}